pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["azfar <azfarmasood22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.19"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
psycopg = "^3.1.19"
psycopg2 = "^2.9.9"
psycopg-binary = "^3.1.19"
psycopg2-binary = "^2.9.9"
tortoise-orm = "^0.21.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.1.3"
python-multipart = "^0.0.9"
aiokafka = "^0.11.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
mypy = "^1.10.0"
ruff = "^0.4.9"
pre-commit = "^3.7.1"
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240327"
coverage = "^7.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
